+++ app/model/ticket.go
@@ -9,12 +9,12 @@ import (
 
 // User struct to describe User object.
 type Ticket struct {
-	ID               uuid.UUID `gorm:"primarykey;type:uuid;default:(uuid_generate_v4())"`
-	CreatedAt        time.Time
-	UpdatedAt        time.Time
-	DeletedAt        gorm.DeletedAt `gorm:"index"`
-	Seat             int
-	Show             string
-	Type             string
-	Price            float
+	ID        uuid.UUID `gorm:"primarykey;type:uuid;default:(uuid_generate_v4())"`
+	CreatedAt time.Time
+	UpdatedAt time.Time
+	DeletedAt gorm.DeletedAt `gorm:"index"`
+	Seat      int
+	Show      string
+	Type      string
+	Price     float64
 }
